+++ data_files/temoa_schema_v3_1.sql
@@ -928,7 +928,7 @@ CREATE TABLE IF NOT EXISTS TimeSeasonSequential
     seas_seq TEXT,
     season TEXT
         REFERENCES TimeSeason (season),
-    count NUMERIC NOT NULL,
+    count REAL NOT NULL,
     notes TEXT,
     PRIMARY KEY (period, sequence, seas_seq, season),
     CHECK (count > 0)
